requirements.txt
python-dotenv==1.0.1
discord==2.3.2
discord.py==2.4.0
pymongo==4.10.1